+++ client/src/components/Header/components/HeaderMenu/components/MenuContact/index.js
@@ -1,0 +1,1 @@
+export { default } from "./MenuContact";
